social_media_api/api/serializers/like_serializers.py

from rest_framework import serializers
from api.models import Like

class LikeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Like
        fields = ['id','user','post','created_at']
        read_only_fields = ['created_at']
