social_media_api/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
drf-yasg
psycopg2-binary
pytest
pytest-django